requirements.txt
telethon>=1.24.0
python-telegram-bot>=20.0
flask>=2.0.0
python-dotenv>=0.19.0